requirements.txt
amqp==5.3.1
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
cssselect2==0.8.0
Django==5.2.4
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.58.5
h11==0.16.0
html5lib==1.1
httptools==0.6.4
idna==3.10
jwt==1.4.0
kombu==5.5.4
lxml==6.0.0
msgpack==1.1.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.8.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.2
requests==2.32.4
rest-framework-simplejwt==0.0.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
